+++ docs/whats_new.md
@@ -1,5 +1,14 @@
 # What's New
 
+## unreleased
+
+* Move known model hard-wired model times into the class so they are refreshed each time the library is read.
+
+## v0.10.1 (January 30, 2025)
+
+* Added built-in way to create plots for simulation using OpenDrift. Details available in {ref}`plots`.
+* User can now input a location to both save and read the interpolator, which avoids using the built-in cache location.
+
 ## v0.9.6 (November 15, 2024)
 
 * made caching directory creation and saving to cache optional with input option `use_cache`.